TrafficFlow/services/data_generator/requirements.txt
kafka-python==2.0.2
numpy==1.26.4
pandas==2.1.4
